refactor(ansi): extract color math out of AnsiOutput into ansi-color.js
`.svelte` files aren't type-checked by tsgo or reachable from bench/
tests, so this color/contrast logic had zero coverage. Moving it to
a plain module makes it testable and benchable; added
bench/ansi.bench.ts as the first baseline (autoContrast costs ~2-3x
more than a flat foreground, now visible instead of implicit).
